server-toxipred/pyproject.toml
[project]
name = "toxipred-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",

  # --- ML stack (kept compatible with NumPy 1.26) ---
  "numpy==1.26.4",
  "scikit-learn==1.4.2",
  "joblib==1.3.2",
  "xgboost==2.0.3",
  "dill>=0.3.8",

  # --- Chemistry ---
  "rdkit-pypi==2023.9.6",   # wheels built for NumPy 1.x

  # --- Queueing / runtime ---
  "celery==5.3.6",
  "redis>=5.0.0",
  "pydantic>=2.7.0",
]
